+++ packages/ai-jsx/src/lib/openai.tsx
@@ -448,14 +448,10 @@ export async function* OpenAIChatModel(
  * @param size The size of the image to generate. Defaults to `512x512`.
  * @returns URL(s) to the generated image, wrapped in {@link Image} component(s).
  */
-export async function DalleImageGen(
+export async function* DalleImageGen(
   { numSamples = 1, size = '512x512', children }: ImageGenPropsWithChildren,
   { render, getContext, logger }: AI.ComponentContext
 ) {
-  const prompt = await render(children);
-
-  const openai = getContext(openAiClientContext);
-
   let sizeEnum;
   switch (size) {
     case '256x256':
@@ -474,6 +470,19 @@ export async function DalleImageGen(
         'user'
       );
   }
+
+  // Consider emitting http://via.placeholder.com/256x256 instead.
+  yield (
+    <Image
+      url={`http://via.placeholder.com/${size}`}
+      prompt="placeholder while real results renderes"
+      modelName="placeholder.com"
+    />
+  );
+
+  const prompt = await render(children);
+
+  const openai = getContext(openAiClientContext);
 
   const imageRequest = {
     prompt,