+++ packages/nextjs-demo/src/app/recipe/api/route.tsx
@@ -14,6 +14,11 @@ const {
   RecipeInstructionListItem,
   RecipeTitle,
 } = RecipeMap;
+import fs from 'fs';
+import path from 'path';
+
+// Flip this flag to use a fixture response. This makes it easier to iterate on the UI.
+const useFixture = false;
 
 export async function POST(request: NextRequest) {
   const { topic } = await request.json();
@@ -24,9 +29,33 @@ export async function POST(request: NextRequest) {
       <UserMessage>Give me a recipe for {topic}.</UserMessage>
     </ChatCompletion>
   );
+
+  // This is an intentional constant flag.
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+  if (useFixture) {
+    const textEncoder = new TextEncoder();
+    const fakeStream = fs.readFileSync(path.join(process.cwd(), 'src', 'app', 'recipe', 'api', 'fixture.txt'), 'utf-8');
+    return new Response(
+      new ReadableStream({
+        start(controller) {
+          controller.enqueue(textEncoder.encode(fakeStream));
+          controller.close();
+        },
+      })
+    );
+  }
+
   return AI.toReactStream(
     RecipeMap,
     <>
+      <ChatCompletion>
+        <SystemMessage>
+          The user will ask for you a recipe. Tell them you'd be happy to do that. Respond in 1-3 sentences. Do not ask
+          a question. Do not give any specific details about the type of recipe you'll return, aside from mentioning the
+          user's topic.
+        </SystemMessage>
+        <UserMessage>I'd like a recipe about {topic}</UserMessage>
+      </ChatCompletion>
       <ImageGen size="256x256">
         <ChatCompletion>
           <UserMessage>Summarize the following recipe into a two sentence description: {recipe}</UserMessage>